feat/water-2644 -- code refactored to follow modules/{area}/ controller and routes -- KPI reporting service to get data from DB pool
